Added user authentication in create edit and delete get

+++ routes/delete.js
@@ -2,6 +2,10 @@ const phoneBookSchema = require('../models/phoneBookSchema');
 
 module.exports = {
     get: (req, res) => {
+        if(!req.session.user){
+            req.session.message='Please login first to delete!';
+            return res.redirect('/user/login');
+        }
         let id = req.params.id;
         phoneBookSchema.findByIdAndDelete(id)
             .then(() => {
